+++ .env.example
@@ -1,11 +1,11 @@
 APP_URL=
+APP_PORT=
 
 DB_NAME=todo
 DB_USER=
 DB_PASS=
 DB_HOST=
 DB_DIALECT=
-PORT=
 
 NODE_ENV=development
