+++ bin/www
@@ -1,90 +1,49 @@
 #!/usr/bin/env node
 
-/**
- * Module dependencies.
- */
-
-import app from '../app';
-var debug = require('debug')('express-todo-api:server');
-import http from 'http';
-import dotenv from 'dotenv';
-
-/**
- * Get port from environment and store in Express.
- */
-
 // .env 사용 설정
+import dotenv from "dotenv";
 dotenv.config();
 
-const port = normalizePort(process.env.PORT || '3000');
-app.set('port', port);
+import http from "http";
+import app from "../src/app.js";
+import debug from "debug";
+import db from "../src/models/index.js";
+
+import redisClient from "../src/config/redis.js";
+
+// 포트 설정
+const port = process.env.APP_PORT || 3000;
+app.set("port", port);
 
 /**
  * Create HTTP server.
  */
+const server = http.createServer(app);
 
-const  server = http.createServer(app);
+// debug 로거 설정
+const debugLogger = debug("express-todo-api:server");
 
-/**
- * Listen on provided port, on all network interfaces.
- */
+// 데이터베이스 연결, 서버 시작
+db.sequelize
+  .sync()
+  .then(() => {
+    console.log("Database connected successfully");
+  })
+  .catch((err) => {
+    console.error("Unable to connect to the database:", err);
+  });
 
-server.listen(port);
-server.on('error', onError);
-server.on('listening', onListening);
+// Redis 연결 확인
+redisClient
+  .connect()
+  .then(() => {
+    console.log("Connected to Redis");
+  })
+  .catch((err) => {
+    console.error("Failed to connect Redis", err);
+  });
 
-/**
- * Normalize a port into a number, string, or false.
- */
-
-function normalizePort(val) {
-  var port = parseInt(val, 10);
-
-  if (isNaN(port)) {
-    // named pipe
-    return val;
-  }
-
-  if (port >= 0) {
-    // port number
-    return port;
-  }
-
-  return false;
-}
-
-/**
- * Event listener for HTTP server "error" event.
- */
-
-function onError(error) {
-  if (error.syscall !== 'listen') {
-    throw error;
-  }
-
-  const bind = typeof port === 'string' ? 'Pipe ' + port : 'Port ' + port;
-
-  // handle specific listen errors with friendly messages
-  switch (error.code) {
-    case 'EACCES':
-      console.error(bind + ' requires elevated privileges');
-      process.exit(1);
-      break;
-    case 'EADDRINUSE':
-      console.error(bind + ' is already in use');
-      process.exit(1);
-      break;
-    default:
-      throw error;
-  }
-}
-
-/**
- * Event listener for HTTP server "listening" event.
- */
-
-function onListening() {
-  const addr = server.address();
-  const bind = typeof addr === 'string' ? 'pipe ' + addr : 'port ' + addr.port;
-  debug('Listening on ' + bind);
-}
+// 서버 시작
+server.listen(port, () => {
+  debugLogger(`Server is running on port ${port}`);
+});
