--- config/config.js
@@ -1,28 +1,0 @@
-//db:migrate 용 config 파일
-import dotenv from 'dotenv';
-
-dotenv.config();
-
-export default {
-  development: {
-    username: process.env.DB_USER || 'root',
-    password: process.env.DB_PASS || null,
-    database: process.env.DB_NAME || 'database_development',
-    host: process.env.DB_HOST || '127.0.0.1',
-    dialect: process.env.DB_DIALECT || 'mysql',
-  },
-  test: {
-    username: process.env.DB_USER || 'root',
-    password: process.env.DB_PASS || null,
-    database: process.env.DB_NAME || 'database_test',
-    host: process.env.DB_HOST || '127.0.0.1',
-    dialect: process.env.DB_DIALECT || 'mysql',
-  },
-  production: {
-    username: process.env.DB_USER || 'root',
-    password: process.env.DB_PASS || null,
-    database: process.env.DB_NAME || 'database_production',
-    host: process.env.DB_HOST || '127.0.0.1',
-    dialect: process.env.DB_DIALECT || 'mysql',
-  },
-};
