+++ package.json
@@ -5,11 +5,12 @@
   "type": "module",
   "scripts": {
     "start": "set DEBUG=myapp:* && node ./bin/www",
-    "dev": "set DEBUG=myapp:* && nodemon app.js"
+    "dev": "set DEBUG=myapp:* && nodemon ./bin/www"
   },
   "dependencies": {
     "argon2": "^0.41.1",
     "cookie-parser": "~1.4.4",
+    "cors": "^2.8.5",
     "debug": "~2.6.9",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
