--- routes/index.js
@@ -1,9 +1,0 @@
-import express from "express";
-const router = express.Router();
-
-/* GET home page. */
-router.get("/", function (req, res) {
-  res.render("index", { title: "Express" });
-});
-
-export default router;
