--- routes/tasks.js
@@ -1,8 +1,0 @@
-import express from "express";
-const router = express.Router();
-import * as taskController from "../controllers/taskController.js";
-
-router.get("/", taskController.getTaskList);
-router.get("/", taskController.getTask);
-
-export default router;
