+++ src/config/keys/private.pem
@@ -1,0 +1,5 @@
+-----BEGIN EC PRIVATE KEY-----
+MHcCAQEEIHKyOs78WXdlXHzbGnPitQ3COAwZP5CwDtU388Swu0tRoAoGCCqGSM49
+AwEHoUQDQgAEfee73J3+AWwK5peRu0lUoFSDn6mlM+pohAt1LruTWEPvxjTvbBis
+ttMgKKD25PqlMm3JbIBnsXMH0mt/Qbp6Sg==
+-----END EC PRIVATE KEY-----
