+++ src/config/keys/public.pem
@@ -1,0 +1,4 @@
+-----BEGIN PUBLIC KEY-----
+MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEfee73J3+AWwK5peRu0lUoFSDn6ml
+M+pohAt1LruTWEPvxjTvbBisttMgKKD25PqlMm3JbIBnsXMH0mt/Qbp6Sg==
+-----END PUBLIC KEY-----
