+++ src/controllers/taskController.js
@@ -1,0 +1,11 @@
+// import db from "../models/index.js";
+
+const getTaskList = async (req, res) => {
+  res.status(200).json({});
+};
+
+const getTask = async (req, res) => {
+  res.status(200).json({});
+};
+
+export { getTaskList, getTask };
